SpotLight class created, bugs

+++ src/PointLight.cpp
@@ -5,8 +5,8 @@ PointLight::PointLight(GLfloat _red, GLfloat _green, GLfloat _blue, GLfloat _amb
 
 PointLight::~PointLight() = default;
 
-void PointLight::useLight(GLfloat _ambientIntensityLocation, GLfloat _ambienColourLocation, GLfloat _diffuseIntensityLocation,
-        GLfloat _positionLocation, GLfloat _constantLocation, GLfloat _linearLocation, GLfloat _exponentLocation) const {
+void PointLight::useLight(GLuint _ambientIntensityLocation, GLuint _ambienColourLocation, GLuint _diffuseIntensityLocation,
+                          GLuint _positionLocation, GLuint _constantLocation, GLuint _linearLocation, GLuint _exponentLocation) const {
     // glUniform — Specify the value of a uniform variable for the current program object
     glUniform3f(_ambienColourLocation, colour.x, colour.y, colour.z);
     glUniform1f(_ambientIntensityLocation, ambientIntensity);
